Fix a few various bugs
There was a bug where all methods define with `_temp_f` would use what
ever the last value of action is. This is due to the late binding of the
action variable. Instead, there needs to be an action key word input to
force the evaluation to occur upfront.
https://stackoverflow.com/questions/3431676/creating-functions-in-a-loop

Also the run method needs to be able to accept key word arguments