Run the command over ssh
The command is echoed into a bash script. `plumbum` requires the first
command argument to be on the users PATH. This would require splitting
the command on the first argument and then generating the plumbum
components individually.

So instead, the command is `echo` into `bash`. This way only the `echo`
and `bash` plumbum components are required.