Add the `batch-id` filter and None guard
The `batch-id` filter was missing, opps. This meant that a job could be
found for basically all the nodes.

However as `batch-id` and `node` form a key to the Job database, exactly
one (or None) jobs will be returned. Thus there needs to be a guard
detecting if no jobs where found.