Make the context passing more explicit
I was trying to be smart and magical, forgetting that this is python not
ruby. The context passing should be explicitly called from the `batch`
module.

This means the `batch` can not be stored within the context as it will
no longer be passed into `Action` (Because python decorators). Instead
it is now passed into the `run_batch` function directly.