Make `open` a globbing command
The `open` command is a plugin command similar to `batch`. It however
looks within the `open` tools directory.

It also saves the node input into the context.

As `open` is a reserved word in `python`, a `open_command` method has
been defined instead. It is then renamed to be `open` and then applies
the decorator manually.