Create a Job object in the `open` command
The `Job` object abstracts away running a command on a node. For this
reason, it can be used within the `open` command. ATM it only prints the
command it would run.

Please note that the `open` command does not save anything to the `db`
and thus no `Session` is required.