Create a ssh connection with `fabric`
`fabric` is the new ssh library that is going to replace `plumbum`.
`fabric` contains more features including interactive sessions.

However it first needs to replace the current `plumbum` code. This
commit causes the `ssh` connection to be created and checked using
`fabric`. The remaining code will be replaced in future commits