Pass the static `Config` into the `actions`
During the `sandbox` commands can be reran multiple times. This meant
that the same `batch` object would be used for all the runs. This is
obviously not ideal.

Instead the static `Config` object is passed in instead. This will allow
multiple `Batch` objects to be created with the same `Config` object