Re-order the table columns
The exit code should be more prominent in the table as it is typically
what the admin wants to see. The date however should be closer to the
end.

The `Last` term has been removed from the headers to reduce the table
width. It can be safely be assumed that this is the last run
information.