Create a local session for each Job
Eventually each Job will be ran within its own thread (as sessions
should not be used across threads).

ATM the "scoped_session" (aka `Session`) object returns the same
instance of `session` for `local_session`. This will change when
multi-threaded.

As an object can not be used in multiple sessions, the `batch` object
needs to be merged in to create a local version

https://docs.sqlalchemy.org/en/latest/orm/contextual.html#unitofwork-contextual