Run the job in a new thread
This is the first step at making the process run in parallel. A
dedicated `run_greenlet` method now runs the job in a separate thread.

On the first context switch, it will need to start the thread. However
this is skipped on future calls.

ATM the thread is `join` before exiting, this will be expanded on in
future commits