Ensure the Sessions are removed in the correct thread
Objects MUST not be shared between threads otherwise the DBMS will have
an aneurysm.

The safest cause of action is to make sure that `Session.remove()` is
called in each thread. This will close the session local to the threads
scope. This prevents the db connection finalizer from trying to close
the session from a different thread