Run the threads in parallel
Up to a maximum of 10 threads can be ran in parallel. This is done by
maintaining a separate list of `active_threads`.

When a thread is started, it is removed from the `threads` list and
placed in `active_threads`. It is then removed from active_threads when
it dies.

`greenlet` has been removed as this can be implemented solely with
python threads