Print the unique Job#id in the `run tool` output
Now that `view result` accepts the `JOB_ID` as its input, it should be
printed during the `run tool` output.

This does cause a few headaches as the `JOB ID` doesn't get assigned
until the `Job` is saved in the db. Currently this is being resolved by
saving the record twice