Only save the job once normally
Previously the job would always be saved twice, firstly to get the job
id and then as part of the `finally`. The `finally` was required because
it ensured the job is saved.

However if the job exited normally, this would cause it to save twice,
increasing the contention on the database.

Instead, the job is saved after it is ran, and thus it gets an ID. All
the remaining tasks need to be ran within a session BUT do not need to
be saved. Thus the `commit` can be removed from the `finally` block.

However the Job is always commited in the event of an error (see
`except` block).