Correctly determine the additional_namespace
Originally all commands where within the `batch` or `open` namespace.
This meant the first namespace could be safely ignored. However when the
command tree was collapsed together, this was not updated.

This meant the first BUT now important namespace was still being
skipped. Due to amount of code that is dependent on the original
constants, a new `TOOL_DIR` constant has been defined.

Also the namespaces should be separated by spaces not `/`