Save the jobs upfront with the Batch
By saving the jobs upfront with the batch, it assigns a Job ID upfront.
This ensures the ID's are sequential even if they take different
amounts of time to run.

It also means the Job objects share an ID between threads. This is
important to stop multiple entries occurring when separate threads write
to the db