Default all new Jobs to an error state
Currently if a Job is queued but not ran, an empty entry is created in
the database. It can be assumed this is an error condition.

As such a new `-4` error code has been added. It denotes an
`Abandoned Job Error`. Abandoned jobs have been created but there is no
record of it ever being ran.