Rename the `shared_job_data_table`
It isn't a particularly good idea to pass a complicated query object
(such as `job_data`) around. If the query was to change, then so will
the rest of the code.

However as the code code is working, it doesn't need to be immediately
replaced. For the time being, the name has been updated so it matches
the variable naming.