Implement a `hashify_all` Config function
This function will convert the `Configs` into a hash. The hash is nested
based on the command name/namespaces. This will allow the config data to
be used with external libraries.

The leaf hashes will contain the payload for each Config. The keys of
this hash are specified by the `kwargs`. `kwargs` also support function
inputs which are evaluated with the Config object. This allows the
values to vary between tools.