Generalise the hashify method
The `subcommand_key` is set to 'commands' b/c that is what the command
generation expects. This is a reasonable default behaviour, but might
not always be the case going forward.

As such, the top level `kwargs` have been broken into `command`,
`group`, and `subcommand_key`. The `subcommand_key` sets the key used to
specify the subcommands. The `command` hash is used to generate the
command leafs.

A third `group` hash will eventually be used to modify the group level
hashes (in addition to the subcommands)