Start rebuilding `run tool`
The current CLI setup is incompatible with how `generate_command` works.

Using `generate_command` removes the ability to pass a variable through
the context. TBF this is a poor design anyway. Instead the options will
be added to to the end of the command.

`generate_command` also passes a different set of inputs into the
command: `callstack`, `argv`, and `options`. The command still needs to
be reworked to use these arguments