Build the `ssh` connection on a kill switch
It is not possible to kill a running thread in python (well not in a
good way).

Instead the solution for killing a running thread is to have another
monitor thread do it for you (IKR?).

So there is now a `threading.Event` on each `Job`. The thread that calls
the `run` method spins up a worker thread and waits for the
`kill_connection_event` to be triggered.

In turn, when the `worker` thread is finished it triggers the
`kill_connection_event` which then terminates the ssh connection.

This `kill_connection_event` can also be triggered externally, hence how
the jobs will be terminated.