Add a `kill` function to the Job runner
This function will flag the job to kill the `ssh` connection. This will
cause the connection to terminate after it has been created.

To prevent weirdness from happening, the `job.run` function should not
clear the flag. This prevents issues with the flag being set after `run`
is called BUT before the `connection` is established.