Run the Jobs asynchronously
The Jobs are now run in `async` mode. This means all the Jobs now share
the same thread, but the scheduler can jump between tasks. It does mean
the order the Jobs are returned in is undefined.

B/c the blocking operation (aka `ssh`) is still present, the Jobs will
not run in parallel.