Remove boiler plate code from execute_threaded_batches
A lot of the code was a holder over from having to create threads. Now
only a single `async` function is required. The method needs to be renamed
at somepoint.

Also, all jobs share a single db connection as they are in the same
thread. Thus all the objects can be saved at the end