Switch `status` to use --node and --group
The status command will be able to query multiple nodes/ groups. Thus it
needs to be updated to use `--node` and `--group`. ATM it can't use
`cli_utils` to do this as that relies on
`appliance_cli.command_generation` option  syntax.

Also the remaining code only works for the first node in the list. This
will be updated in a bit