Add try-except block for config creation
Now if FileNotFound error occurs the system will not crash

Also, in the case of the file not being found, a defaultdict will be created for the config instead of the dict created by reading its yaml.
It has the default value for everything 'File Not Found'.
This dict is currently not used for anything as as it stands all queried information comes from the job object but it exists for future utility.

fixes #103