Use the subqueries within the count function
The major purpose of the subqueries is to generate a list Job.ids that
corresponds to the `tool` and `node` inputs.

However because subqueries do not have a `ORM` abstraction, it is hard
to get the `Job` object out of them. There maybe a way to do this, but
I'm not aware of it.

Instead a new query to a job table and `id_query` subquery table is
used. This allows for the `Job` object to be loaded on the condition its
`id` appears in the `id_query`.

The `config` is also included in the `id_query` as this allows for the
`group_by` to work without having to `join` in the `Batches` table.

The `max_func` is include but ignored in the result. The purpose of
this is to ensure the `Job` object returned out of the select
corresponds to the maximum `Job.created_date` (i.e. the most recent job
within the group).