Extract the job_query generation
The `history` commands are being absorbed into `status` so they can
share the same code. This means `job_query` needs to be generalised in a
manner that allows it to poll for all past jobs.

The easiest way to do this is to toggle on the `funcs`. The SQL funcs
require a `group_by` to operate properly. This allows them to be used as
a toggle quite effectively.