Cache the file globbing for configs
The config directories do not need to be re globbed each time a new tool
is created. Instead, the glob is being cached which ensures all tools
have the same configs.

The entire command isn't cached as this would cache a cached output (as
`Config.cache` is also being used)