Create a mixin to handle model reconstruction
A bit of setup is required in order to handle the reconstruction of a
model correctly. SQLAlchemy does not call `_init__` when loading a model
from the db.

This is why a custom `__init__` method is required, otherwise the
additional setup code will not run.

The `orm.reconstructor` flags the method to be called when loading from
the db