Build the reconstructor into the Base model
This means inherited classes do not need to include the mixin. Instead
they can use the reconstructor automagically

The semantics on how this can be achieved are pretty knarly. Please
consult the docs for more information:
https://sqlalchemy-html.readthedocs.io/en/rel_1_0_6/orm/extensions/declarative/api.html

In short, by setting `constructor = None` it actually uses the
constructor defined in the initial version of `Base`

However b/c SQLAlchemy defines it's constructors in an odd manner,
`super` is not available within the initial version of `Base`. Instead
it needs to manually call the `_declarative_constructor`

To special methods have been setup to use the reconstructor:
__load__ which will run on loading from the db only
__init_or_load__ which will run on load and init

Redefining one will not affect the other