Add a ShellVariable database table/model
Shell variables are essentially key value pairs that are associated with
a `Batch`. It isn't a great idea to store lists within a db and as such
each variable will be stored in its own table.

The basic design of ShellVariable has been modelled of Job (w/o the
running code)