Extract the finished method for the Task
Unfortunately the python3.5 version of `Task` doesn't have a means to
determine if a task has finished vs done. Instead the internal `_state`
method needs to be checked.

To increase the maintainability, the method for doing this has been
extracted into `finished`. This will make it easier to update in the
future if need be