Downcase argument names in cmds when cmds are ran
The change so that argument names are downcased in order
to match their values returned as from click (the commit before last)
moved the problem such that there would be miss-match between the
argument names stored by the system and those specified in the tool's
cmd. This resulted in arguments not being used.

This commit searches through a tools cmd on execution to fix this
miss-match by replacing all instances with downcased versions.